renames package to memcached
