+++ memcached-driver/src/test/java/org/jnosql/diana/memcached/key/HazelcastBucketManagerQueryTest.java
@@ -12,13 +12,13 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.diana.hazelcast.key;
+package org.jnosql.diana.memcached.key;
 
 import com.hazelcast.query.Predicate;
 import org.jnosql.diana.api.Value;
 import org.jnosql.diana.api.key.BucketManagerFactory;
-import org.jnosql.diana.hazelcast.key.model.Movie;
-import org.jnosql.diana.hazelcast.key.util.KeyValueEntityManagerFactoryUtils;
+import org.jnosql.diana.memcached.key.model.Movie;
+import org.jnosql.diana.memcached.key.util.KeyValueEntityManagerFactoryUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
