+++ memcached-driver/src/test/java/org/jnosql/diana/memcached/key/QueueTest.java
@@ -13,12 +13,12 @@
  *   Otavio Santana
  */
 
-package org.jnosql.diana.hazelcast.key;
+package org.jnosql.diana.memcached.key;
 
 
 import org.jnosql.diana.api.key.BucketManagerFactory;
-import org.jnosql.diana.hazelcast.key.model.LineBank;
-import org.jnosql.diana.hazelcast.key.util.KeyValueEntityManagerFactoryUtils;
+import org.jnosql.diana.memcached.key.model.LineBank;
+import org.jnosql.diana.memcached.key.util.KeyValueEntityManagerFactoryUtils;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
