+++ memcached-driver/src/test/java/org/jnosql/diana/memcached/key/model/Movie.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.diana.hazelcast.key.model;
+package org.jnosql.diana.memcached.key.model;
 
 import java.io.Serializable;
