+++ memcached-driver/src/test/resources/diana-hazelcast.properties
@@ -14,4 +14,4 @@
 #
 
 hazelcast-instanceName=hazelcast
-#hazelcast-host-1=
+#memcached-host-1=
